requirements.txt
google-auth-oauthlib==1.2.0
PyPDF2==3.0.1
google-api-python-client==2.121.0